Change functions directory
