+++ src/components/Board/index.spec.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 import { render, within, act, fireEvent, waitForElement } from '@testing-library/react'
 import Board from './'
 import { callbacks } from 'react-beautiful-dnd'
-import { moveCard } from './services'
+import { moveCard } from '@services/helpers'
 
 describe('<Board />', () => {
   let subject, onCardDragEnd, onLaneDragEnd, onLaneRemove, onLaneRename, onCardRemove
